requirements.txt
flask==2.1.3
flask-restx==0.5.1
flask-session==0.4.0
python-dotenv==0.21.0
requests==2.28.1
gradio==3.9.1
httpx==0.22.0
